refactor(admin): extract AdminRoleGuard to eliminate duplicate layout logic
